+++ src/App.css
@@ -1,4 +1,3 @@
-/* src/App.css */
 .app-container {
   display: flex;
   flex-direction: column;
