+++ src/pages/NotFound.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const NotFound = ():React.ReactNode => {
+  return <h2>404 - Page Not Found</h2>;
+};
+
+export default NotFound;
